Fix Automatic Quarter Sizing (#726)
* Change approach to forcibly setting column count

This is what we implicitly do already by setting the min quarter width. This commit changes it so that the column count is explicitly set instead of implicitly set.

Note that we also don't need perfect pixel precision because the natural column count at the breakpoint already forms the same number of columns. They aren't affected by the previous breakpoint because of the data-max-quarter-count attribute.

For example, at a screen width of 1999px and 5 quarters, grid template columns is not overwritten and there are naturally 5 quarters. At 2000px, the column count is forcibly set to 5. At some point between 2000 and 2250px (+250 because that's the min quarter width), that's when it'd naturally go to 6 columns, but because we overrode at 2000px, it will not go to 6 if the max number of quarters is < 6.

* Convert media query logic into loop